# dataland-backend/backendOpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Dataland Backend API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/public/"
    }
  ],
...
  "paths": {
...
    "/public/companies": {
      "get": {
        "tags": [
          "company-data-controller"
        ],
        "summary": "Retrieve specific companies by different filters or just all companies from the data store.",
        "description": "Companies identified via the provided company name/identifier are retrieved and filtered bycountryCode, sector and available framework data. Empty/Unspecified filters are ignored.",
        "operationId": "getCompanies",
        "parameters": [
          {
            "name": "searchString",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataTypes",
            "in": "query",
            "required": false,
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataTypeEnum"
              }
            }
          },
          {
            "name": "countryCodes",
            "in": "query",
            "required": false,
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sectors",
            "in": "query",
            "required": false,
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "onlyCompanyNames",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved companies.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoredCompany"
                  }
                }
              }
            }
          },
          "default": {
            "description": "An error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "default-oauth": []
          },
          {
            "default-bearer-auth": []
          },
          {
            "dataland-api-key": []
          }
        ]
      },
      "post": {
        "tags": [
          "company-data-controller"
        ],
        "summary": "Add a new company.",
        "description": "A new company is added using the provided information, the generated company ID is returned.",
        "operationId": "postCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyInformation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully added company.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredCompany"
                }
              }
            }
          },
          "default": {
            "description": "An error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "default-oauth": []
          },
          {
            "default-bearer-auth": []
          },
          {
            "dataland-api-key": []
          }
        ]
      }
    },
...
    "/public/metadata": {
      "get": {
        "tags": [
          "meta-data-controller"
        ],
        "summary": "Search in Dataland for meta info about data.",
        "description": "Meta info about data sets registered by Dataland can be retrieved.",
        "operationId": "getListOfDataMetaInfo",
        "parameters": [
          {
            "name": "companyId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataType",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DataTypeEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved meta info.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataMetaInformation"
                  }
                }
              }
            }
          },
          "default": {
            "description": "An error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "default-oauth": []
          },
          {
            "default-bearer-auth": []
          },
          {
            "dataland-api-key": []
          }
        ]
      }
    },
    "/public/metadata/{dataId}": {
      "get": {
        "tags": [
          "meta-data-controller"
        ],
        "summary": "Look up meta info about a specific data set.",
        "description": "Meta info about a specific data set registered by Dataland and identified by its data ID is retrieved.",
        "operationId": "getDataMetaInfo",
        "parameters": [
          {
            "name": "dataId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved specific meta info.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataMetaInformation"
                }
              }
            }
          },
          "default": {
            "description": "An error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "default-oauth": []
          },
          {
            "default-bearer-auth": []
          },
          {
            "dataland-api-key": []
          }
        ]
      }
    },
...
    "/public/companies/{companyId}": {
      "get": {
        "tags": [
          "company-data-controller"
        ],
        "summary": "Retrieve company information.",
        "description": "Company information behind the given company Id is retrieved.",
        "operationId": "getCompanyById",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved company information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredCompany"
                }
              }
            }
          },
          "default": {
            "description": "An error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "default-oauth": []
          },
          {
            "default-bearer-auth": []
          },
          {
            "dataland-api-key": []
          }
        ]
      }
    },
    "/public/companies/teaser": {
      "get": {
        "tags": [
          "company-data-controller"
        ],
        "summary": "Get the company IDs of the teaser companies.",
        "description": "A list of all company IDs that are currently set as teaser companies (accessible without authentication).",
        "operationId": "getTeaserCompanies",
        "responses": {
          "200": {
            "description": "Successfully returned teaser companies.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "An error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "default-oauth": []
          },
          {
            "default-bearer-auth": []
          },
          {
            "dataland-api-key": []
          }
        ]
      }
    },
    "/public/companies/meta-information": {
      "get": {
        "tags": [
          "company-data-controller"
        ],
        "summary": "Retrieve available distinct values for company search filters",
        "description": "Distinct values for the parameter countryCode and sector are returned",
        "operationId": "getAvailableCompanySearchFilters",
        "responses": {
          "200": {
            "description": "Successfully retrieved values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyAvailableDistinctValues"
                }
              }
            }
          },
          "default": {
            "description": "An error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "default-oauth": []
          },
          {
            "default-bearer-auth": []
          },
          {
            "dataland-api-key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataMetaInformation": {
        "required": [
          "companyId",
          "dataId",
          "dataType",
          "isQualityAssured",
          "uploadTime"
        ],
        "type": "object",
        "properties": {
          "dataId": {
            "type": "string"
          },
          "dataType": {
            "$ref": "#/components/schemas/DataTypeEnum"
          },
          "uploaderUserId": {
            "type": "string"
          },
          "uploadTime": {
            "type": "integer",
            "format": "int64"
          },
          "companyId": {
            "type": "string"
          },
          "isQualityAssured": {
            "type": "string"
          }
        }
      },
...
      "CompanyIdentifier": {
        "required": [
          "identifierType",
          "identifierValue"
        ],
        "type": "object",
        "properties": {
          "identifierType": {
            "type": "string",
            "enum": [
              "Lei",
              "Isin",
              "PermId",
              "Ticker",
              "Duns"
            ]
          },
          "identifierValue": {
            "type": "string"
          }
        }
      },
      "CompanyInformation": {
        "required": [
          "companyName",
          "countryCode",
          "headquarters",
          "identifiers",
          "sector"
        ],
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string"
          },
          "companyAlternativeNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "companyLegalForm": {
            "type": "string"
          },
          "headquarters": {
            "type": "string"
          },
          "headquartersPostalCode": {
            "type": "string"
          },
          "sector": {
            "type": "string"
          },
          "identifiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyIdentifier"
            }
          },
          "countryCode": {
            "type": "string"
          },
          "isTeaserCompany": {
            "type": "boolean"
          }
        }
      },
      "StoredCompany": {
        "required": [
          "companyId",
          "companyInformation",
          "dataRegisteredByDataland"
        ],
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "companyInformation": {
            "$ref": "#/components/schemas/CompanyInformation"
          },
          "dataRegisteredByDataland": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataMetaInformation"
            }
          }
        }
      },
...
      "CompanyAvailableDistinctValues": {
        "type": "object",
        "properties": {
          "countryCodes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sectors": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "DataTypeEnum": {
        "type": "string",
        "enum": [
          "eutaxonomy-financials",
          "eutaxonomy-non-financials",
          "lksg",
          "sfdr",
          "sme"
        ]
      },
      "ErrorDetails": {
        "type": "object",
        "properties": {
          "httpStatus": {
            "type": "number"
          },
          "errorType": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetails"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "dataland-api-key": {
        "type": "apiKey",
        "name": "dataland-api-key",
        "in": "header"
      },
      "default-bearer-auth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer"
      },
      "default-oauth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/keycloak/realms/datalandsecurity/protocol/openid-connect/auth",
            "tokenUrl": "/keycloak/realms/datalandsecurity/protocol/openid-connect/token",
            "scopes": {}
          }
        }
      }
    }
  }
}